The function 'getRegularFunctionParameter' (and then Length etc) now returns by default numerical approximations.
If not, in certain circumstances, we get in a MemoryError.
Example : in picture ContourTgNDivergence [1] we get these points on a
parametric curve, then compute the tangent vector at these points and
check if the length is zero (because we want to normalize). The problem is that the
length is something like
 sqrt(((cos(6891/16384*pi)^2 - sin(6891/16384*pi)^2)*cos(6891/16384*pi) - (cos(6891/16384*pi)*sin(6891/16384*pi) + 1)*sin(6891/16384*pi))^2 + ((cos(6891/16384*pi)*sin(6891/16384*pi) + 1)*cos(6891/16384*pi) + (cos(6891/16384*pi)^2 - sin(6891/16384*pi)^2)*sin(6891/16384*pi))^2)

It takes forever to Sage to check if it is zero or not. And in some
cases, Sage gets in MemoryError.

[1] https://github.com/LaurentClaessens/mazhe