The axes use the 'action_on_pspict' mechanism.
This caused a lot a changes at many places :
(1) first compute the math bounding box of the picture
(2) adapt the default axes, enlarge them and draw them
(3) compute the bounding box of the picture.

In order to not re-compute the bounding boxes of the objects in (3), the
Picture have a 'already_computed_BB' list that records the objects
for which the bounding box is already computed and taken into account.