'PointGraph.__eq__' first check the type.
When computing the BB of a Picture, we check for each element if it is
already in the 'already_computed_BB' list.
This generates tests like
obj1 == obj2
for obj2 in the already computed list. When obj2 is a 'PointGraph', the
equality test often crashes because 'obj1' has no 'x' attribute.
And when it does not crashes, this opens a bug possibility.