(mark graph) Make write only 5 digits in the coordinates of the central point of a mark.
There are too much non-deterministic operations here and there. So
limiting the number of digits at the end of the day is the best way to
avoid non-deterministic '.pstricks' final file.

The last one I found is in `AngleMeasure.measure :`
```
return AngleMeasure(value_degree=self.angleF.degree-self.angleI.degree)
```
In the picture FBTC the difference is
`206.565051177078- 341.565051177078` which leads to problems.